reimplement chunking; near-infinite undo/redo; fixes; regressive commit
no longer ever modify chunks, only replace them.

- as before, chunks define "chains", ie.  circular lists of chunks
- whatever edit op is performed, all overlapping chunks are *replaced*
by new ones representing the new content of the interval + sliced
borders of the two terminal chunks on each side:

	-[ a ]-[ b ] ..  [ c ]-[ d ]-
	  replace ^^^^^^^^^^
	-[ a ]-[L][x] ..  [y][R]-[ d ]-
	  new undo frame: [ b ]-...-[ c ], [L]-[x]-..-[y]-[R]

	-[ a ]-[ b ] ..  [ c ]-[ d ]-
	  cut ^^^^^^^^^^
	-[ a ]-[L]-[R]-[ d ]-
	  new undo frame: [ b ]-...-[ c ], [L]-[R]

	-[ a ]-[ b ] ..  [ c ]-[ d ]-
	  insert ^
	-[ a ]-[L]-[ i ]-[R]-[ c ]-[ d ]-
	  new undo frame: [ b ], [L]-[ i ]-[R]

	same with insert/replace/crop/copy/etc.

- bordering overlapping chunks are sliced into two; the outermost
slices are named L and R resp.  for the left and right border and just
their truncated unmodified prefix/suffix; the entire overlapping chain
is unlinked and replaced by a new chain terminated by L and R; the op
is pushed to an undo stack by just saving the pointers to the "parent"
unsliced left and right border, and pointers to L and R.
- always replacing entire overlapping chunks means that all pointers
in between are always valid and never need repair, that undo/redo
merely substitute one chain for another, and that all operations are
in constant time; the cost in terms of chunks is dwarfed by the audio
buffer size either way regardless of the multiplication of the number
of chunks, and this dispenses with locks and other complicated logic
for snarfing buffers; it also dispenses with the need to ever memcpy
anything, only pointers are passed around.

this resembles or replicates umbraticus' own method in pcmed(1).

also:
- "forget" ops if undoing then rewriting history from an earlier point
- just allocate new undo slots as needed instead of even limiting
them, who cares, negligeable in practice
- synchronous edit after read proc is done reading in new data
- lower i/o buffer size; don't memcpy anything...  for now.
- ui: there is always a valid range selected; upon any left click,
we set an insertion point; upon any right click we set a replacement
range and unset any insertion point; simpler than prior approach
- display total time duration of range

not thoroughly tested and there are regressions -- all much easier
to debug.